+++ csv/csv_parse_stream_test.ts
@@ -2,14 +2,13 @@
 import { CsvParseStream } from "./csv_parse_stream.ts";
 import type { CsvParseStreamOptions } from "./csv_parse_stream.ts";
 import { ERR_QUOTE, ParseError } from "./_io.ts";
-import { readableStreamFromIterable } from "../streams/readable_stream_from_iterable.ts";
 import { readableStreamFromReader } from "../streams/readable_stream_from_reader.ts";
 import {
   assert,
   assertEquals,
   assertRejects,
   assertStringIncludes,
-} from "../testing/asserts.ts";
+} from "../assert/mod.ts";
 import type { AssertTrue, IsExact } from "../testing/types.ts";
 import { fromFileUrl, join } from "../path/mod.ts";
 import { StringReader } from "../io/string_reader.ts";
@@ -42,7 +41,7 @@ Deno.test({
 Deno.test({
   name: "[csv/csv_parse_stream] CsvParseStream with invalid csv",
   fn: async () => {
-    const readable = readableStreamFromIterable([
+    const readable = ReadableStream.from([
       encoder.encode("id,name\n"),
       encoder.encode("\n"),
       encoder.encode("1,foo\n"),