+++ examples/echo_server_test.ts
@@ -3,7 +3,7 @@ import {
   assertEquals,
   assertNotEquals,
   assertStrictEquals,
-} from "../testing/asserts.ts";
+} from "../assert/mod.ts";
 import { BufReader, ReadLineResult } from "../io/buf_reader.ts";
 import { dirname, fromFileUrl } from "../path/mod.ts";
 import { TextLineStream } from "../streams/text_line_stream.ts";