+++ http/testdata/file_server_as_library.ts
@@ -1,6 +1,8 @@
-import { serve } from "../server.ts";
 import { serveFile } from "../file_server.ts";
 
-serve((req) => {
-  return serveFile(req, "./testdata/hello.html");
-}, { port: 8000, onListen: () => console.log("Server running...") });
+Deno.serve(
+  { port: 8000, onListen: () => console.log("Server running...") },
+  (req) => {
+    return serveFile(req, "./testdata/hello.html");
+  },
+);