+++ semver/test_comparator_test.ts
@@ -1,6 +1,6 @@
 // Copyright Isaac Z. Schlueter and Contributors. All rights reserved. ISC license.
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-import { assert } from "../testing/asserts.ts";
+import { assert } from "../assert/mod.ts";
 import { parse } from "./parse.ts";
 import { parseComparator } from "./parse_comparator.ts";
 import { testComparator } from "./test_comparator.ts";